+++ test/assign.test.cc
@@ -9,10 +9,10 @@ TEST_CASE("Tensor Assignment", "[int]") {
   auto tensor_1 = Tensor<int32_t, 4>({1, 2, 3, 4});
 
   // Initialize values
-  for (uint32_t i = 1; i <= tensor_1.dimension(1); ++i)
-    for (uint32_t j = 1; j <= tensor_1.dimension(2); ++j)
-      for (uint32_t k = 1; k <= tensor_1.dimension(3); ++k)
-        for (uint32_t l = 1; l <= tensor_1.dimension(4); ++l) 
+  for (size_t i = 1; i <= tensor_1.dimension(1); ++i)
+    for (size_t j = 1; j <= tensor_1.dimension(2); ++j)
+      for (size_t k = 1; k <= tensor_1.dimension(3); ++k)
+        for (size_t l = 1; l <= tensor_1.dimension(4); ++l) 
           tensor_1(i, j, k, l) = 1000 * i + 100 * j + 10 * k + l;
 
   /*     --------------------------   */
@@ -21,74 +21,74 @@ TEST_CASE("Tensor Assignment", "[int]") {
 
     /* same type */
     Tensor<int32_t, 2> t2({3, 4});
-    for (uint32_t i = 1; i <= t2.dimension(1); ++i)
-      for (uint32_t j = 1; j <= t2.dimension(2); ++j) 
+    for (size_t i = 1; i <= t2.dimension(1); ++i)
+      for (size_t j = 1; j <= t2.dimension(2); ++j) 
         t2(i, j) = tensor_1(1, 1, i, j) - 1000;
 
-    for (uint32_t i = 1; i <= t2.dimension(1); ++i)
-      for (uint32_t j = 1; j <= t2.dimension(2); ++j) 
+    for (size_t i = 1; i <= t2.dimension(1); ++i)
+      for (size_t j = 1; j <= t2.dimension(2); ++j) 
         REQUIRE(t2(i, j) == (int)(100 + 10 * i + j));
 
     tensor_1(1, 1) = t2;
 
-    for (uint32_t i = 1; i <= tensor_1.dimension(3); ++i)
-      for (uint32_t j = 1; j <= tensor_1.dimension(4); ++j)
+    for (size_t i = 1; i <= tensor_1.dimension(3); ++i)
+      for (size_t j = 1; j <= tensor_1.dimension(4); ++j)
         REQUIRE(tensor_1(1, 1, i, j) == (int)(100 + 10 * i + j));
 
-    for (uint32_t i = 1; i <= tensor_1.dimension(3); ++i)
-      for (uint32_t j = 1; j <= tensor_1.dimension(4); ++j)
+    for (size_t i = 1; i <= tensor_1.dimension(3); ++i)
+      for (size_t j = 1; j <= tensor_1.dimension(4); ++j)
         REQUIRE(tensor_1(1, 2, i, j) == (int)(1000 + 200 + 10 * i + j));
 
     /* different type */
     Tensor<double, 4> t3({1, 2, 3, 4});
-    for (uint32_t i = 1; i <= t3.dimension(1); ++i)
-      for (uint32_t j = 1; j <= t3.dimension(2); ++j)
-        for (uint32_t k = 1; k <= t3.dimension(3); ++k)
-          for (uint32_t l = 1; l <= t3.dimension(4); ++l)
+    for (size_t i = 1; i <= t3.dimension(1); ++i)
+      for (size_t j = 1; j <= t3.dimension(2); ++j)
+        for (size_t k = 1; k <= t3.dimension(3); ++k)
+          for (size_t l = 1; l <= t3.dimension(4); ++l)
             t3(i, j, k, l) = -1 * (int)(i + 10 * j + 100 * k + 1000 * l);
 
     tensor_1 = t3;
 
-    for (uint32_t i = 1; i <= t3.dimension(1); ++i)
-      for (uint32_t j = 1; j <= t3.dimension(2); ++j)
-        for (uint32_t k = 1; k <= t3.dimension(3); ++k)
-          for (uint32_t l = 1; l <= t3.dimension(4); ++l)
+    for (size_t i = 1; i <= t3.dimension(1); ++i)
+      for (size_t j = 1; j <= t3.dimension(2); ++j)
+        for (size_t k = 1; k <= t3.dimension(3); ++k)
+          for (size_t l = 1; l <= t3.dimension(4); ++l)
             REQUIRE(tensor_1(i, j, k, l) == -1 * (int)(i + 10 * j + 100 * k + 1000 * l));
   }
 
   SECTION("Assigning Integral Different-Typed Tensors to Tensors") {
     Tensor<long double, 2> t2({3, 4});
-    for (uint32_t i = 1; i <= t2.dimension(1); ++i)
-      for (uint32_t j = 1; j <= t2.dimension(2); ++j)
+    for (size_t i = 1; i <= t2.dimension(1); ++i)
+      for (size_t j = 1; j <= t2.dimension(2); ++j)
         t2(i, j) = tensor_1(1, 1, i, j) - 1000;
 
-    for (uint32_t i = 1; i <= t2.dimension(1); ++i)
-      for (uint32_t j = 1; j <= t2.dimension(2); ++j)
+    for (size_t i = 1; i <= t2.dimension(1); ++i)
+      for (size_t j = 1; j <= t2.dimension(2); ++j)
         REQUIRE(t2(i, j) == tensor_1(1, 1, i, j) - 1000);
 
     tensor_1(1, 1) = t2;
 
-    for (uint32_t i = 1; i <= tensor_1.dimension(3); ++i)
-      for (uint32_t j = 1; j <= tensor_1.dimension(4); ++j)
+    for (size_t i = 1; i <= tensor_1.dimension(3); ++i)
+      for (size_t j = 1; j <= tensor_1.dimension(4); ++j)
         REQUIRE(tensor_1(1, 1, i, j) == (int)(100 + 10 * i + j));
 
 
-    for (uint32_t i = 1; i <= tensor_1.dimension(3); ++i)
-      for (uint32_t j = 1; j <= tensor_1.dimension(4); ++j)
+    for (size_t i = 1; i <= tensor_1.dimension(3); ++i)
+      for (size_t j = 1; j <= tensor_1.dimension(4); ++j)
         REQUIRE(tensor_1(1, 2, i, j) == (int)(1000 + 200 + 10 * i + j));
 
     Tensor<int64_t, 4> t3({1, 2, 3, 4});
-    for (uint32_t i = 1; i <= t3.dimension(1); ++i)
-      for (uint32_t j = 1; j <= t3.dimension(2); ++j)
-        for (uint32_t k = 1; k <= t3.dimension(3); ++k)
-          for (uint32_t l = 1; l <= t3.dimension(4); ++l)
+    for (size_t i = 1; i <= t3.dimension(1); ++i)
+      for (size_t j = 1; j <= t3.dimension(2); ++j)
+        for (size_t k = 1; k <= t3.dimension(3); ++k)
+          for (size_t l = 1; l <= t3.dimension(4); ++l)
             t3(i, j, k, l) = -1 * (int)(i + 10 * j + 100 * k + 1000 * l);
 
     tensor_1 = t3;
-    for (uint32_t i = 1; i <= t3.dimension(1); ++i)
-      for (uint32_t j = 1; j <= t3.dimension(2); ++j)
-        for (uint32_t k = 1; k <= t3.dimension(3); ++k)
-          for (uint32_t l = 1; l <= t3.dimension(4); ++l)
+    for (size_t i = 1; i <= t3.dimension(1); ++i)
+      for (size_t j = 1; j <= t3.dimension(2); ++j)
+        for (size_t k = 1; k <= t3.dimension(3); ++k)
+          for (size_t l = 1; l <= t3.dimension(4); ++l)
             REQUIRE(tensor_1(i, j, k, l) == -1 * (int)(i + 10 * j + 100 * k + 1000 * l));
   }
 
@@ -148,10 +148,10 @@ TEST_CASE("Scalar Assignment", "[int]") {
   Tensor<int32_t> stensor_1{};
 
   // Initialize values
-  for (uint32_t i = 1; i <= tensor_1.dimension(1); ++i)
-    for (uint32_t j = 1; j <= tensor_1.dimension(2); ++j)
-      for (uint32_t k = 1; k <= tensor_1.dimension(3); ++k)
-        for (uint32_t l = 1; l <= tensor_1.dimension(4); ++l) {
+  for (size_t i = 1; i <= tensor_1.dimension(1); ++i)
+    for (size_t j = 1; j <= tensor_1.dimension(2); ++j)
+      for (size_t k = 1; k <= tensor_1.dimension(3); ++k)
+        for (size_t l = 1; l <= tensor_1.dimension(4); ++l) {
           tensor_1(i, j, k, l) = 1000 * i + 100 * j + 10 * k + l;
         }
 
@@ -161,73 +161,73 @@ TEST_CASE("Scalar Assignment", "[int]") {
 
   SECTION("Assigning Tensors to Tensors") {
     Tensor<int32_t, 2> t2({3, 4});
-    for (uint32_t i = 1; i <= t2.dimension(1); ++i)
-      for (uint32_t j = 1; j <= t2.dimension(2); ++j) 
+    for (size_t i = 1; i <= t2.dimension(1); ++i)
+      for (size_t j = 1; j <= t2.dimension(2); ++j) 
         t2(i, j) = tensor_1(1, 1, i, j) - 1000;
 
-    for (uint32_t i = 1; i <= t2.dimension(1); ++i)
-      for (uint32_t j = 1; j <= t2.dimension(2); ++j) 
+    for (size_t i = 1; i <= t2.dimension(1); ++i)
+      for (size_t j = 1; j <= t2.dimension(2); ++j) 
         REQUIRE(t2(i, j) == (int)(100 + 10 * i + j));
 
     tensor_1(1, 1) = t2;
 
-    for (uint32_t i = 1; i <= tensor_1.dimension(3); ++i)
-      for (uint32_t j = 1; j <= tensor_1.dimension(4); ++j)
+    for (size_t i = 1; i <= tensor_1.dimension(3); ++i)
+      for (size_t j = 1; j <= tensor_1.dimension(4); ++j)
         REQUIRE(tensor_1(1, 1, i, j) == (int)(100 + 10 * i + j));
 
 
-    for (uint32_t i = 1; i <= tensor_1.dimension(3); ++i)
-      for (uint32_t j = 1; j <= tensor_1.dimension(4); ++j)
+    for (size_t i = 1; i <= tensor_1.dimension(3); ++i)
+      for (size_t j = 1; j <= tensor_1.dimension(4); ++j)
         REQUIRE(tensor_1(1, 2, i, j) == (int)(1000 + 200 + 10 * i + j));
 
     Tensor<int32_t, 4> t3({1, 2, 3, 4});
-    for (uint32_t i = 1; i <= t3.dimension(1); ++i)
-      for (uint32_t j = 1; j <= t3.dimension(2); ++j)
-        for (uint32_t k = 1; k <= t3.dimension(3); ++k)
-          for (uint32_t l = 1; l <= t3.dimension(4); ++l)
+    for (size_t i = 1; i <= t3.dimension(1); ++i)
+      for (size_t j = 1; j <= t3.dimension(2); ++j)
+        for (size_t k = 1; k <= t3.dimension(3); ++k)
+          for (size_t l = 1; l <= t3.dimension(4); ++l)
             t3(i, j, k, l) = -1 * (int)(i + 10 * j + 100 * k + 1000 * l);
 
     tensor_1 = t3;
-    for (uint32_t i = 1; i <= t3.dimension(1); ++i)
-      for (uint32_t j = 1; j <= t3.dimension(2); ++j)
-        for (uint32_t k = 1; k <= t3.dimension(3); ++k)
-          for (uint32_t l = 1; l <= t3.dimension(4); ++l)
+    for (size_t i = 1; i <= t3.dimension(1); ++i)
+      for (size_t j = 1; j <= t3.dimension(2); ++j)
+        for (size_t k = 1; k <= t3.dimension(3); ++k)
+          for (size_t l = 1; l <= t3.dimension(4); ++l)
             REQUIRE(tensor_1(i, j, k, l) == -1 * (int)(i + 10 * j + 100 * k + 1000 * l));
   }
 
   SECTION("Assigning Integral Different-Typed Tensors to Tensors") {
     Tensor<long double, 2> t2({3, 4});
-    for (uint32_t i = 1; i <= t2.dimension(1); ++i)
-      for (uint32_t j = 1; j <= t2.dimension(2); ++j)
+    for (size_t i = 1; i <= t2.dimension(1); ++i)
+      for (size_t j = 1; j <= t2.dimension(2); ++j)
         t2(i, j) = tensor_1(1, 1, i, j) - 1000;
 
-    for (uint32_t i = 1; i <= t2.dimension(1); ++i)
-      for (uint32_t j = 1; j <= t2.dimension(2); ++j)
+    for (size_t i = 1; i <= t2.dimension(1); ++i)
+      for (size_t j = 1; j <= t2.dimension(2); ++j)
         REQUIRE(t2(i, j) == tensor_1(1, 1, i, j) - 1000);
 
     tensor_1(1, 1) = t2;
 
-    for (uint32_t i = 1; i <= tensor_1.dimension(3); ++i)
-      for (uint32_t j = 1; j <= tensor_1.dimension(4); ++j)
+    for (size_t i = 1; i <= tensor_1.dimension(3); ++i)
+      for (size_t j = 1; j <= tensor_1.dimension(4); ++j)
         REQUIRE(tensor_1(1, 1, i, j) == (int)(100 + 10 * i + j));
 
 
-    for (uint32_t i = 1; i <= tensor_1.dimension(3); ++i)
-      for (uint32_t j = 1; j <= tensor_1.dimension(4); ++j)
+    for (size_t i = 1; i <= tensor_1.dimension(3); ++i)
+      for (size_t j = 1; j <= tensor_1.dimension(4); ++j)
         REQUIRE(tensor_1(1, 2, i, j) == (int)(1000 + 200 + 10 * i + j));
 
     Tensor<int64_t, 4> t3({1, 2, 3, 4});
-    for (uint32_t i = 1; i <= t3.dimension(1); ++i)
-      for (uint32_t j = 1; j <= t3.dimension(2); ++j)
-        for (uint32_t k = 1; k <= t3.dimension(3); ++k)
-          for (uint32_t l = 1; l <= t3.dimension(4); ++l)
+    for (size_t i = 1; i <= t3.dimension(1); ++i)
+      for (size_t j = 1; j <= t3.dimension(2); ++j)
+        for (size_t k = 1; k <= t3.dimension(3); ++k)
+          for (size_t l = 1; l <= t3.dimension(4); ++l)
             t3(i, j, k, l) = -1 * (int)(i + 10 * j + 100 * k + 1000 * l);
 
     tensor_1 = t3;
-    for (uint32_t i = 1; i <= t3.dimension(1); ++i)
-      for (uint32_t j = 1; j <= t3.dimension(2); ++j)
-        for (uint32_t k = 1; k <= t3.dimension(3); ++k)
-          for (uint32_t l = 1; l <= t3.dimension(4); ++l)
+    for (size_t i = 1; i <= t3.dimension(1); ++i)
+      for (size_t j = 1; j <= t3.dimension(2); ++j)
+        for (size_t k = 1; k <= t3.dimension(3); ++k)
+          for (size_t l = 1; l <= t3.dimension(4); ++l)
             REQUIRE(tensor_1(i, j, k, l) == -1 * (int)(i + 10 * j + 100 * k + 1000 * l));
   }
